+++ src/locale/index.ts
@@ -1,0 +1,8 @@
+export type { Locale, xLocale } from './types';
+export type { LocaleContextProps } from './context';
+export { LocaleContextKey } from './context';
+export { default as useLocale } from './useLocale';
+export type { LocaleComponentName } from './useLocale';
+export { default as LocaleProvider } from './LocaleProvider.vue';
+export { default as zhCN } from './zh_CN';
+export { default as enUS } from './en_US';